09.2 delete requests fix

+++ lessons/views.py
@@ -45,7 +45,7 @@ def lesson_request(request):
                 teacher=teacher
             )
             lessonRequest.save()
-            return redirect('lesson_request')
+            return redirect('student_home')
     else:
         form = LessonRequestForm()
     return render(request, 'lesson_request.html', {'form': form})
@@ -170,14 +170,11 @@ def admin_requests(request):
     return render(request, 'admin_lesson_list.html', {'data': lesson_request_data})
 
 @login_required
+@only_students
 def show_requests(request):
-    try:
-        user = request.user
-        lesson_requests = LessonRequest.objects.filter(author=user)
-    except ObjectDoesNotExist:
-        return redirect('home')
-    else:
-        return render(request, 'show_requests.html', {'user': user, 'lesson_requests': lesson_requests})
+    user = request.user
+    lesson_requests = LessonRequest.objects.filter(author=user)
+    return render(request, 'show_requests.html', {'user': user, 'lesson_requests': lesson_requests})
 
 '''allow users to change passwords'''
 @login_required
@@ -218,6 +215,7 @@ def profile(request):
     return render(request, 'profile.html', {'form': form})
 
 @login_required
+@only_students
 def edit_requests(request, lesson_id):
     try:
         current_lesson = LessonRequest.objects.get(id=lesson_id)
@@ -233,11 +231,17 @@ def edit_requests(request, lesson_id):
             form = EditForm(instance=current_lesson)
         return render(request, 'edit_requests.html', {'form': form, 'lesson_id': lesson_id})
 
+@login_required
+@only_students
 def delete_requests(request, lesson_id):
     try:
         current_lesson = LessonRequest.objects.get(id=lesson_id)
     except ObjectDoesNotExist:
         return redirect('show_requests')
     else:
-        current_lesson.delete()
-        return redirect('show_requests')
+        user = request.user
+        if current_lesson.author != user:
+            return redirect('home')
+        else:
+            current_lesson.delete()
+            return redirect('show_requests')
